requirements.txt
yt-dlp
moviepy
opencv-python-headless
Pillow
instagrapi
argparse
diffusers
transformers
torch